+++ pinax/news/specs.py
@@ -1,0 +1,14 @@
+from imagekit import ImageSpec
+from pilkit.processors import SmartResize, ResizeToFit
+
+
+class ImageThumbnail(ImageSpec):
+    processors = [ResizeToFit(width=168, height=None)]
+    format = "JPEG"
+    options = {"quality": 75}
+
+
+class SecondaryImageThumbnail(ImageSpec):
+    processors = [SmartResize(405, 250)]
+    format = "JPEG"
+    options = {"quality": 75}