+++ docs/changelog.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## 1.1.0
+
+* added support for secondary images
+* added support for customized image sizing
+
 ## 1.0.0
 
 * added docs and tests and wired up CI